Clean up the misuse of gitignore
The operating model is now to whitelist directories and files at the
root level. Every subsequent level will then use gitignore as a
blacklist, which is is a common use case. This will prevent cases where
we have files that are ignored when we actually want to add it to repo.