[git] Rebase only on ancestor
We use `git rb` to rebase and squash changes. But we use `git pull` when
we want to merge changes from upstream.
This lets `rb` be used often and is less disruptive than having to deal
with merge conflicts when you're not wanting to pull in changes.

Also remove tags from `git ahead`. You only care about which commits
have diverged, it's often redundant information to include refs.