[fish] Correct PATH to work in all circumstances
We incorrectly had status is-login/is-interactive checks to set the
PATH. This works most of the time because the tools from these
additional paths were interactive tools.  But this is incorrect because
the PATH is not correct when subshell or script is invoked.

Since these values are static and not exported, unconditionally set them
for every shell invocation.