[git] Change push.default to current
The default is simple, change to current. `simple` only work when the
upstream branch is the same name. We cannot use this to track against
origin/main.

This makes `git push` push the current branch to a matching branch on
the remote. The remote is the same as upstream branch. This setup allows
pushing to the correct remote and not hardcoded to origin. It can be
individually overridden per-branch.