[fish] Implement history expansion with plugin-expand
We've always had `!` to trigger history expansion. It's like ctrl-r but
for inserting into the middle of command, such as for a while loop.
Now, this supports a query. So in the middle of writing a while loop,
you begin to type "foo --bar" and realize it's better to insert from
history. You can do so without having to discard partial typing.