[vim] Add firenvim firefox extension
Add firefox firenvim extension settings. This extension is configured in
init.vim and the extension communicates with a headless neovim instance
to "do the needful". Hence why we have all the GUI logic here.

Besides minimizing the UI, due to limited real estate in viewport, this
also has shrinking and growing of multiline input.