[nvim] Make copilotchat a floating window
A first-class window was too much real estate for copilot chat.
Floating window is better usecase, since we will focus and can use a
large pane size, even crossing existing windows.

And it turns out to be lightweight workflow to close the preview window
and restore it as needed, instead of persistently on.